Refactored code and folder structure
Moved project files into folders.
Did changes to how mongo is closed, this gets updated in the next commit.
