+++ package.json
@@ -2,7 +2,7 @@
   "name": "ad-export-sync",
   "version": "1.0.0",
   "description": "",
-  "main": "client.js",
+  "main": "./client/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
